feat: enable Lombok builder pattern and fix MapStruct DTO mappings
Added @Builder, @NoArgsConstructor, and @AllArgsConstructor annotations to response DTOs for compatibility with MapStruct code generation.
Resolved compilation errors due to constructor mismatches and unmapped fields.
Configured JWT token provider with required properties via application config.
Refactored bean structure to eliminate circular dependency between InvoiceController and InvoiceServiceImpl.

Project builds successfully and launches without runtime errors. ✅

+++ src/main/resources/db/migration/V1__create_initial_schema.sql
@@ -1,13 +1,13 @@
 -- Tabela de Sementes
 CREATE TABLE seeds (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     species VARCHAR(100) NOT NULL,
     cultivar VARCHAR(100) NOT NULL
 );
 
 -- Tabela de Clientes
 CREATE TABLE clients (
-     id SERIAL PRIMARY KEY,
+     id BIGSERIAL PRIMARY KEY,
      number VARCHAR(20) NOT NULL,
      name VARCHAR(100) NOT NULL,
      email VARCHAR(100),
@@ -18,7 +18,7 @@ CREATE TABLE clients (
 
 -- Tabela de Usuários
 CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     email VARCHAR(100) NOT NULL UNIQUE,
     password VARCHAR(100) NOT NULL,
     name VARCHAR(100) NOT NULL,
@@ -31,10 +31,10 @@ CREATE TABLE users (
 
 -- Tabela de Notas Fiscais
 CREATE TABLE invoices (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     invoice_number VARCHAR(50) NOT NULL,
     producer_name VARCHAR(100) NOT NULL,
-    seed_id INTEGER REFERENCES seeds(id),
+    seed_id BIGINT REFERENCES seeds(id),
     total_kg DECIMAL(10,2) NOT NULL,
     operation_type VARCHAR(50) NOT NULL, -- REPACKAGING, TRANSFER
     auth_number VARCHAR(50),
@@ -49,47 +49,47 @@ CREATE TABLE invoices (
 
 -- Tabela de Lotes
 CREATE TABLE lots (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     lot_number VARCHAR(50) NOT NULL UNIQUE,
     lot_type VARCHAR(20) NOT NULL, -- INTERNAL_SALE, EXPORT
-    seed_id INTEGER REFERENCES seeds(id),
+    seed_id BIGINT REFERENCES seeds(id),
     seed_type VARCHAR(30) NOT NULL, -- COATED, GRAPHITE, COATED_GRAPHITE, CONVENTIONAL
     category VARCHAR(10) NOT NULL, -- S1, S2
     bag_weight DECIMAL(10,2) NOT NULL, -- 1, 5, 10, 15, 20, 25, 600, 800, 1000
     balance DECIMAL(10,2) NOT NULL,
     analysis_bulletin VARCHAR(50),
     bulletin_date DATE,
-    invoice_id INTEGER REFERENCES invoices(id),
+    invoice_id BIGINT REFERENCES invoices(id),
     bag_type VARCHAR(30),
     validity_date DATE,
     seed_score INTEGER,
-    user_id INTEGER NOT NULL REFERENCES users(id),
+    user_id BIGINT NOT NULL REFERENCES users(id),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 -- Tabela de Baixas de Lotes
 CREATE TABLE lot_withdrawals (
-    id SERIAL PRIMARY KEY,
-    lot_id INTEGER REFERENCES lots(id),
+    id BIGSERIAL PRIMARY KEY,
+    lot_id BIGINT REFERENCES lots(id),
     invoice_number VARCHAR(50),
     quantity DECIMAL(10,2) NOT NULL,
     withdrawal_date DATE NOT NULL,
     state VARCHAR(2) NOT NULL,
-    user_id INTEGER NOT NULL REFERENCES users(id),
-    client_id INTEGER REFERENCES clients(id),
+    user_id BIGINT NOT NULL REFERENCES users(id),
+    client_id BIGINT REFERENCES clients(id),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 -- Tabela de Reservas de Lotes
 CREATE TABLE lot_reservations (
-    id SERIAL PRIMARY KEY,
-    lot_id INTEGER REFERENCES lots(id),
+    id BIGSERIAL PRIMARY KEY,
+    lot_id BIGINT REFERENCES lots(id),
     quantity DECIMAL(10,2) NOT NULL,
     reservation_date DATE NOT NULL,
     status VARCHAR(20) NOT NULL, -- PENDING, CONFIRMED, CANCELLED
-    client_id INTEGER REFERENCES clients(id),
-    user_id INTEGER NOT NULL REFERENCES users(id),
+    client_id BIGINT REFERENCES clients(id),
+    user_id BIGINT NOT NULL REFERENCES users(id),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
